### 2018-08-29 #246 @khufkens
No rush, I promised both @sckott and @seantuck12 to push this back to rOpensci back in June but the summer got in the way.
