### 2018-09-05 #246 @khufkens
I fixed most of the notes above, but haven't pushed them. Keep me posted on progress.
