### 2018-09-06 1003668
small changes using seq_len() and replacing stray = with <-
