### 2018-09-08 22969c0
server string now formats version separately, for future updates
