### 2018-09-08 a8386ab
additional tests for site_id calls
