### 2018-09-14 #246 @etiennebr
Ok, I won't worry about the CRAN version; are the fix pushed to the github master branch?
