### 2018-09-14 #246 @khufkens
Yes, everything is in the master branch. Larger additions or refactoring is done on a separate one (but not applicable now).
