### 2018-10-07 a29dd36
replace NULL with missing
