### 2018-10-27 fa521d6
use missing instead of NULL for empty parameters
