### 2018-10-28 12e1c25
include zzz.R options for server and api version, remove (stray) NULL parameters, update DESCRIPTION and NEWS file
