+++ NEWS.md
@@ -1,3 +1,10 @@
+## MODISTools 1.1.0
+
+* ropensci code review
+* replaced NULL parameters with missing()
+* output a tidy data frame (tibble)
+* include the coordinates of the pixel(s) as an sf object
+
 ## MODISTools 1.0.0
 
 * ropensci refactoring
