+++ R/zzz.R
@@ -3,8 +3,7 @@
 .onLoad <- function(libname, pkgname) {
   # set server options
   options(list(
-    mt_server = "https://modis.ornl.gov/rst/api/",
-    mt_api_version = "v1"
+    mt_server = "https://modis.ornl.gov/rst/api/v1/"
     )
   )
 }
