+++ tests/testthat/test_main_functions.r
@@ -1,163 +1,109 @@
 # main unit tests
 
 test_that("test mt_products()",{
-
-  # test listing all products
-  products <- try(mt_products())
-
-  # see if any of the runs failed
-  check = !inherits(products, "try-error")
-
-  # check if no error occured
-  expect_true(check)
+  expect_is(mt_products(), "data.frame")
 })
 
 test_that("test mt_sites()",{
-
-  # test listing all products
-  products <- try(mt_sites())
-
-  # see if any of the runs failed
-  check = !inherits(products, "try-error")
-
-  # check if no error occured
-  expect_true(check)
+  expect_is(mt_sites(), "data.frame")
 })
 
 test_that("test mt_bands()",{
-
-  # test listing all products
-  bands <- try(mt_bands(product = "MOD11A2"))
-  bands_err <- try(mt_bands(product = "MOD11A6"))
-
-  # see if any of the runs failed
-  check = !inherits(bands, "try-error") &
-    inherits(bands_err, "try-error")
-
-  # check if no error occured
-  expect_true(check)
+  expect_is(mt_bands(product = "MOD11A2"), "data.frame")
+  expect_error(mt_bands(product = "MOD11A6"))
 })
 
-
 test_that("test mt_dates()",{
+  expect_is(mt_dates(product = "MOD11A2",
+                          lat = 40,
+                          lon = -110),
+            "data.frame")
 
-  # test listing all products
-  dates <- try(mt_dates(product = "MOD11A2",
+  expect_error(mt_dates(product = "MOD11A6",
                           lat = 40,
                           lon = -110))
 
-  dates_err <- try(mt_dates(product = "MOD11A6",
-                          lat = 40,
-                          lon = -110))
-
-  dates_coord_err <- try(mt_dates(product = "MOD11A6",
+  expect_error(mt_dates(product = "MOD11A6",
                               lat = NULL,
                               lon = -110))
 
-  # grab a random site_id for a pre-processed site
-  id <- mt_sites()$siteid[1]
+  expect_is(mt_dates(product = "MOD11A2",
+                                site_id = mt_sites()$siteid[1]),
+            "data.frame")
 
-  dates_site_id <- try(mt_dates(product = "MOD11A2",
-                                site_id = id))
-
-  # see if any of the runs failed
-  check = !inherits(dates, "try-error") &
-    !inherits(dates_site_id, "try-error") &
-    inherits(dates_err, "try-error") &
-    inherits(dates_coord_err, "try-error")
-
-  # check if no error occured
-  expect_true(check)
+  expect_error(mt_dates(product = "MOD11A2",
+                     site_id = "test"))
 })
 
 test_that("test mt_subset()",{
 
-  # download data
-  subset = try(mt_subset(product = "MOD11A2",
+  expect_is(mt_subset(product = "MOD11A2",
                     lat = 40,
                     lon = -110,
                     band = "LST_Day_1km",
                     start = "2004-01-01",
-                    end = "2004-03-31"))
+                    end = "2004-03-31"),
+            "data.frame")
 
-  # download data
-  subset_no_loc = try(mt_subset(product = "MOD11A2",
+  expect_error(mt_subset(product = "MOD11A2",
                          lat = 40,
                          band = "LST_Day_1km",
                          start = "2004-01-01",
                          end = "2004-03-31"))
 
-  # download data
-  subset_site_id = try(mt_subset(product = "MOD11A2",
+  expect_is(mt_subset(product = "MOD11A2",
                                  site_id = "us_tennessee_neon_ornl",
                                  band = "LST_Day_1km",
                                  start = "2004-01-01",
-                                 end = "2004-03-31"))
[…]
