+++ vignettes/modistools-vignette.Rmd
@@ -24,7 +24,7 @@ The MODISTools package has as goal to facilitate the interface between R and the
 
 ### Listing products / bands / dates
 
-In order to assess which products are available, which product bands are provided and which temporal range is covered one has to list these ancillary data. All these options can be queried using the mt_*() functions.
+In order to assess which products are available, which product bands are provided and which temporal range is covered one has to list these ancillary data. All these options can be queried using the `mt_*()` functions.
 
 To list all available products use the mt_products() function.
 
@@ -36,29 +36,29 @@ head(products)
 To list all available bands for a given product use the mt_bands() function.
 
 ```{r eval = TRUE}
-bands <- mt_bands(product = "MOD11A2")
+bands <- mt_bands(product = "MOD13Q1")
 head(bands)
 ```
 
 To list all available dates (temporal coverage) for a given product and location use the mt_dates() function.
 
 ```{r eval = TRUE}
-dates <- mt_dates(product = "MOD11A2", lat = 42, lon = -110)
+dates <- mt_dates(product = "MOD13Q1", lat = 42, lon = -110)
 head(dates)
 ```
 
 ### Downloading MODIS time series
 
-Once you decide on which data to download using the above functions you can use these parameters to download a time series using the mt_subset() function. The below query downloads MOD11A2 based daytime lands surface temperatures. The location is named 'testsite' and only covers the first six months of 2004. The output is saved to a variable called subset in the R workspace (as defined by the parameter internal = TRUE, when set to FALSE the data is written to file).
+Once you decide on which data to download using the above functions you can use these parameters to download a time series using the `mt_subset()` function. The below query downloads MOD11A2 based daytime lands surface temperatures. The location is named 'testsite' and only covers the first six months of 2004. The output is saved to a variable called subset in the R workspace (as defined by the parameter internal = TRUE, when set to FALSE the data is written to file).
 
 ```{r eval = TRUE}
 # download data
-subset <- mt_subset(product = "MOD11A2",
-                    lat = 40,
-                    lon = -110,
-                    band = "LST_Day_1km",
+subset <- mt_subset(product = "MOD13Q1",
+                    lat = 42.534171,
+                    lon = -72.179003,
+                    band = "250m_16_days_NDVI",
                     start = "2004-01-01",
-                    end = "2004-06-01",
+                    end = "2008-12-30",
                     km_lr = 0,
                     km_ab = 0,
                     site_name = "testsite",
@@ -67,27 +67,11 @@ head(subset)
 
 ```
 
-The output format is a nested list with the true data provided as a *tidy* data frame, as shown above. When witten to a csv, when the parameter 'internal' is set to FALSE, the same information is retained. Data can be read back into the same format with the included mt_read() function (see below).
+The output format is a *tidy* data frame, as shown above. When witten to a csv with the parameter `internal = FALSE` this will result in a flat file on disk.
 
-Note that when a a region is defined using km_lr and km_ab multiple pixels might be returned. These are indexed using the 'pixel' column in the data frame containing the time series data. The remote sensing values are listed in the 'data' column. When no band is specified all bands of a given product are returned, be mindful of the fact that different bands might require different multipliers to represent their true values. 
+Note that when a a region is defined using km_lr and km_ab multiple pixels might be returned. These are indexed using the `pixel` column in the data frame containing the time series data. The remote sensing values are listed in the `value` column. When no band is specified all bands of a given product are returned, be mindful of the fact that different bands might require different multipliers to represent their true values. 
 
[…]
