+++ R/coordinate_conversion.R
@@ -1,0 +1,118 @@
+#' Convert sinusoidal coordinates to lat / lon
+#'
+#' @param x sinusoidal x coordinate (vector)
+#' @param y sinusoidal y coordinate (vector)
+#' @seealso [latlon_to_square()] [square_to_poly()]
+#' @keywords MODIS Land Products Subsets, products, meta-data
+#' @export
+#' @importFrom magrittr %>%
+#' @examples
+#'
+#' \donttest{
+#' # Download some test data
+#' subset <- mt_subset(product = "MOD11A2",
+#'                         lat = 40,
+#'                         lon = -110,
+#'                         band = "LST_Day_1km",
+#'                         start = "2004-01-01",
+#'                         end = "2004-03-31")
+#'
+#' # convert sinusoidal to lat / lon
+#' subset <- subset %>%
+#'     mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) %>%
+#'     unnest()
+#'}
+
+sin_to_latlon <- function(x, y){
+  coords <- sf::st_as_sf(x = data.frame(as.numeric(x),
+                                        as.numeric(y),
+                                        stringsAsFactors = FALSE),
+                     coords = c("as.numeric.x.", "as.numeric.y."),
+                     crs = "+proj=sinu +a=6371007.181 +b=6371007.181 +units=m")
+
+  # reproject coordinates
+  coords <- sf::st_transform(coords, "+init=epsg:4326")
+
+  # unravel the sf dataframe into a normal one
+  coords <- as.data.frame(do.call("rbind", lapply(coords$geometry, unlist)))
+  colnames(coords) <- c("longitude_ll", "latitude_ll")
+
+  # return the labelled dataframe
+  return(coords)
+}
+
+#' Convert lat / lon coordinates to square bounding box
+#'
+#' @param x sinusoidal x coordinate (vector)
+#' @param y sinusoidal y coordinate (vector)
+#' @param res resolution of a single pixel
+#' @param nrow number of rows in the MODISTools call
+#' @param ncol number of columns in the MODISTools call
+#' @seealso [sin_to_latlon()] [square_to_poly()]
+#' @keywords MODIS Land Products Subsets, products, meta-data
+#' @export
+#' @importFrom magrittr %>%
+#' @examples
+#'
+#' \donttest{
+#' # Download some test data
+#' subset <- mt_subset(product = "MOD11A2",
+#'                         lat = 40,
+#'                         lon = -110,
+#'                         band = "LST_Day_1km",
+#'                         start = "2004-01-01",
+#'                         end = "2004-03-31")
+#'
+#' # convert sinusoidal to lat / lon
+#' subset <- subset %>%
+#'     mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) %>%
+#'     unnest()
+#'
+#'
+#'}
+
+latlon_to_square <- function(x = 0, y = 0, res = 500, nrow = 1, ncol = 1 ){
+  if(missing(res)){stop("no resolution provied")}
+  m_deg = (0.00001/1.1132)
+  trans <- c(
+    0, 0,
+    0, 1,
+    1, 1,
+    1, 0,
+    0, 0
+  )
+  matrix(c(-res * ncol * m_deg, res * nrow * m_deg) * trans + c(x, y), 5, 2, byrow = TRUE)
+}
+
+#' Convert square coordinates to sf bounding box polygon
+#'
+#' @param m matrix as generated by [latlon_to_square()]
+#' @seealso [latlon_to_square()] [sin_to_latlon()]
+#' @keywords MODIS Land Products Subsets, products, meta-data
+#' @export
+#' @importFrom magrittr %>%
[…]
