+++ man/latlon_to_square.Rd
@@ -1,0 +1,50 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/coordinate_conversion.R
+\name{latlon_to_square}
+\alias{latlon_to_square}
+\title{Convert lat / lon coordinates to square bounding box}
+\usage{
+latlon_to_square(x = 0, y = 0, res = 500, nrow = 1, ncol = 1)
+}
+\arguments{
+\item{x}{sinusoidal x coordinate (vector)}
+
+\item{y}{sinusoidal y coordinate (vector)}
+
+\item{res}{resolution of a single pixel}
+
+\item{nrow}{number of rows in the MODISTools call}
+
+\item{ncol}{number of columns in the MODISTools call}
+}
+\description{
+Convert lat / lon coordinates to square bounding box
+}
+\examples{
+
+\donttest{
+# Download some test data
+subset <- mt_subset(product = "MOD11A2",
+                        lat = 40,
+                        lon = -110,
+                        band = "LST_Day_1km",
+                        start = "2004-01-01",
+                        end = "2004-03-31")
+
+# convert sinusoidal to lat / lon
+subset <- subset \%>\%
+    mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) \%>\%
+    unnest()
+
+
+}
+}
+\seealso{
+[sin_to_latlon()] [square_to_poly()]
+}
+\keyword{Land}
+\keyword{MODIS}
+\keyword{Products}
+\keyword{Subsets,}
+\keyword{meta-data}
+\keyword{products,}
