### 2018-10-31 8e2ff32
coordinate conversions docs
