+++ R/coordinate_conversion.R
@@ -19,11 +19,11 @@
 #'
 #' # convert sinusoidal to lat / lon
 #' subset <- subset %>%
-#'     mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) %>%
+#'     mutate(result = map2(xllcorner, yllcorner, sin_to_ll)) %>%
 #'     unnest()
 #'}
 
-sin_to_latlon <- function(x, y){
+sin_to_ll <- function(x, y){
   coords <- sf::st_as_sf(x = data.frame(as.numeric(x),
                                         as.numeric(y),
                                         stringsAsFactors = FALSE),
@@ -41,14 +41,14 @@ sin_to_latlon <- function(x, y){
   return(coords)
 }
 
-#' Convert lat / lon coordinates to square bounding box
+#' Convert lat / lon coordinates to sf bounding box polygon
 #'
-#' @param x sinusoidal x coordinate (vector)
-#' @param y sinusoidal y coordinate (vector)
-#' @param res resolution of a single pixel
-#' @param nrow number of rows in the MODISTools call
-#' @param ncol number of columns in the MODISTools call
-#' @seealso [sin_to_latlon()] [square_to_poly()]
+#' @param lat latitude generated by [sin_to_ll()]
+#' @param lon longitude generated by [sin_to_ll()]
+#' @param cell_size cell size provided by MODISTools
+#' @param nrows cell size provided by MODISTools
+#' @param ncols cell size provided by MODISTools
+#' @seealso [sin_to_ll()]
 #' @keywords MODIS Land Products Subsets, products, meta-data
 #' @export
 #' @importFrom magrittr %>%
@@ -65,15 +65,29 @@ sin_to_latlon <- function(x, y){
 #'
 #' # convert sinusoidal to lat / lon
 #' subset <- subset %>%
-#'     mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) %>%
+#'     mutate(result = map2(xllcorner, yllcorner, sin_to_ll)) %>%
 #'     unnest()
 #'
 #'
 #'}
 
-latlon_to_square <- function(x = 0, y = 0, res = 500, nrow = 1, ncol = 1 ){
-  if(missing(res)){stop("no resolution provied")}
+ll_to_bb <- function(
+  lat,
+  lon,
+  cell_size,
+  nrow,
+  ncol
+  ){
+
+  # check parameters
+  if(missing(res)){
+    stop("missing parameter, all parameters are required")
+    }
+
+  # meters to a degree
   m_deg = (0.00001/1.1132)
+
+  # transform matrix
   trans <- c(
     0, 0,
     0, 1,
@@ -81,36 +95,12 @@ latlon_to_square <- function(x = 0, y = 0, res = 500, nrow = 1, ncol = 1 ){
     1, 0,
     0, 0
   )
-  matrix(c(-res * ncol * m_deg, res * nrow * m_deg) * trans + c(x, y), 5, 2, byrow = TRUE)
-}
 
-#' Convert square coordinates to sf bounding box polygon
-#'
-#' @param m matrix as generated by [latlon_to_square()]
-#' @seealso [latlon_to_square()] [sin_to_latlon()]
-#' @keywords MODIS Land Products Subsets, products, meta-data
-#' @export
-#' @importFrom magrittr %>%
-#' @examples
-#'
-#' \donttest{
-#' # Download some test data
-#' subset <- mt_subset(product = "MOD11A2",
-#'                         lat = 40,
-#'                         lon = -110,
-#'                         band = "LST_Day_1km",
-#'                         start = "2004-01-01",
-#'                         end = "2004-03-31")
[…]
