+++ man/sin_to_ll.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/coordinate_conversion.R
-\name{sin_to_latlon}
-\alias{sin_to_latlon}
+\name{sin_to_ll}
+\alias{sin_to_ll}
 \title{Convert sinusoidal coordinates to lat / lon}
 \usage{
-sin_to_latlon(x, y)
+sin_to_ll(x, y)
 }
 \arguments{
 \item{x}{sinusoidal x coordinate (vector)}
@@ -27,7 +27,7 @@ subset <- mt_subset(product = "MOD11A2",
 
 # convert sinusoidal to lat / lon
 subset <- subset \%>\%
-    mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) \%>\%
+    mutate(result = map2(xllcorner, yllcorner, sin_to_ll)) \%>\%
     unnest()
 }
 }
