+++ man/ll_to_bb.Rd
@@ -1,16 +1,24 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/coordinate_conversion.R
-\name{square_to_poly}
-\alias{square_to_poly}
-\title{Convert square coordinates to sf bounding box polygon}
+\name{ll_to_bb}
+\alias{ll_to_bb}
+\title{Convert lat / lon coordinates to sf bounding box polygon}
 \usage{
-square_to_poly(m)
+ll_to_bb(lat, lon, cell_size, nrow, ncol)
 }
 \arguments{
-\item{m}{matrix as generated by [latlon_to_square()]}
+\item{lat}{latitude generated by [sin_to_ll()]}
+
+\item{lon}{longitude generated by [sin_to_ll()]}
+
+\item{cell_size}{cell size provided by MODISTools}
+
+\item{nrows}{cell size provided by MODISTools}
+
+\item{ncols}{cell size provided by MODISTools}
 }
 \description{
-Convert square coordinates to sf bounding box polygon
+Convert lat / lon coordinates to sf bounding box polygon
 }
 \examples{
 
@@ -25,14 +33,14 @@ subset <- mt_subset(product = "MOD11A2",
 
 # convert sinusoidal to lat / lon
 subset <- subset \%>\%
-    mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) \%>\%
+    mutate(result = map2(xllcorner, yllcorner, sin_to_ll)) \%>\%
     unnest()
 
 
 }
 }
 \seealso{
-[latlon_to_square()] [sin_to_latlon()]
+[sin_to_ll()]
 }
 \keyword{Land}
 \keyword{MODIS}
