+++ NAMESPACE
@@ -1,12 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
-export(latlon_to_square)
+export(ll_to_bb)
 export(mt_bands)
 export(mt_batch_subset)
 export(mt_dates)
 export(mt_products)
 export(mt_sites)
 export(mt_subset)
-export(sin_to_latlon)
-export(square_to_poly)
+export(sin_to_ll)
 importFrom(magrittr,"%>%")
