### 2018-10-31 e5238e7
coordinate conversion docs
