### 2018-10-31 bb9b708
simplified package dependencies, not relying on the tidyverse
