+++ NAMESPACE
@@ -8,4 +8,3 @@ export(mt_products)
 export(mt_sites)
 export(mt_subset)
 export(sin_to_ll)
-importFrom(magrittr,"%>%")
