### 2018-10-31 6350d57
update coordinate conversion parameters
