+++ man/ll_to_bb.Rd
@@ -4,7 +4,7 @@
 \alias{ll_to_bb}
 \title{Convert lat / lon coordinates to sf bounding box polygon}
 \usage{
-ll_to_bb(lat, lon, cell_size, nrow, ncol)
+ll_to_bb(lat, lon, cell_size, nrows, ncols)
 }
 \arguments{
 \item{lat}{latitude generated by [sin_to_ll()]}
