### 2018-10-31 eafb847
batch subset progress bar suppression
