+++ .travis.yml
@@ -4,7 +4,9 @@ cache: packages
 warnings_are_errors: true
 
 before_install:
-  - sudo apt-get install -y libudunits2-dev
+  - sudo apt-get install --yes udunits-bin libudunits2-dev
+  - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev libgdal1-dev
+  - sudo apt-get install --yes libnetcdf-dev netcdf-bin
 
 r:
   - release
