+++ README.md
@@ -1,16 +1,23 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/khufkens/MODISTools.svg)](https://travis-ci.org/khufkens/MODISTools) [![codecov](https://codecov.io/gh/khufkens/MODISTools/branch/master/graph/badge.svg)](https://codecov.io/gh/khufkens/MODISTools) ![Status](https://www.r-pkg.org/badges/version/MODISTools) ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/MODISTools)
+
+[![Build
+Status](https://travis-ci.org/khufkens/MODISTools.svg)](https://travis-ci.org/khufkens/MODISTools)
+[![codecov](https://codecov.io/gh/khufkens/MODISTools/branch/master/graph/badge.svg)](https://codecov.io/gh/khufkens/MODISTools)
+![Status](https://www.r-pkg.org/badges/version/MODISTools)
+![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/MODISTools)
 <a href="https://www.buymeacoffee.com/H2wlgqCLO" target="_blank"><img src="https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png" alt="Buy Me A Coffee" height="21px" ></a>
 <a href="https://liberapay.com/khufkens/donate"><img alt="Donate using Liberapay" src="https://liberapay.com/assets/widgets/donate.svg" height="21px"></a>
 
-MODISTools
-==========
+# MODISTools
 
-Programmatic interface to the ['MODIS Land Products Subsets' web services](https://modis.ornl.gov/data/modis_webservice.html). Allows for easy downloads of ['MODIS'](http://modis.gsfc.nasa.gov/) time series directly to your R workspace or your computer. When using the package please cite the manuscript as referenced below.
+Programmatic interface to the [‘MODIS Land Products Subsets’ web
+services](https://modis.ornl.gov/data/modis_webservice.html). Allows for
+easy downloads of [‘MODIS’](http://modis.gsfc.nasa.gov/) time series
+directly to your R workspace or your computer. When using the package
+please cite the manuscript as referenced below.
 
-Installation
-------------
+## Installation
 
 ### stable release
 
@@ -23,7 +30,8 @@ library("MODISTools")
 
 ### development release
 
-To install the development releases of the package run the following commands:
+To install the development releases of the package run the following
+commands:
 
 ``` r
 if(!require(devtools)){install.package("devtools")}
@@ -31,7 +39,8 @@ devtools::install_github("khufkens/MODISTools")
 library("MODISTools")
 ```
 
-Vignettes are not rendered by default, if you want to include additional documentation please use:
+Vignettes are not rendered by default, if you want to include additional
+documentation please use:
 
 ``` r
 if(!require(devtools)){install.package("devtools")}
@@ -39,79 +48,37 @@ devtools::install_github("khufkens/MODISTools", build_vignettes = TRUE)
 library("MODISTools")
 ```
 
-Use
----
+## Use
 
 ### Downloading MODIS time series
 
-To extract a time series of modis data for a given location and its direct environment use the mt\_subset() function.
+To extract a time series of modis data for a given location and its
+direct environment use the mt\_subset() function.
 
-<details><summary>detailed parameter description (click to expand)</summary>
+<details>
+
+<summary>detailed parameter description (click to
[…]
