### 2018-10-31 7d74fdd
updating zzz with function
