+++ R/zzz.R
@@ -1,9 +1,3 @@
 # load default server names
 
-.onLoad <- function(libname, pkgname) {
-  # set server options
-  options(list(
-    mt_server = "https://modis.ornl.gov/rst/api/v1/"
-    )
-  )
-}
+mt_server <- function() 'https://modis.ornl.gov/rst/api/v1/'
