### 2018-10-31 9eabce5
updating docs and references
