+++ tests/testthat/test_main_functions.r
@@ -1,5 +1,5 @@
-# main unit tests
-
+# test functions without parameters
+# only can fail upon server error
 test_that("test mt_products()",{
   expect_is(mt_products(), "data.frame")
 })
@@ -14,58 +14,132 @@ test_that("test mt_bands()",{
 })
 
 test_that("test mt_dates()",{
-  expect_is(mt_dates(product = "MOD11A2",
-                          lat = 40,
-                          lon = -110),
-            "data.frame")
 
-  expect_error(mt_dates(product = "MOD11A6",
-                          lat = 40,
-                          lon = -110))
+  # check dates
+  expect_is(
+    mt_dates(
+      product = "MOD11A2",
+      lat = 40,
+      lon = -110
+    ),
+  "data.frame"
+  )
 
-  expect_error(mt_dates(product = "MOD11A6",
-                              lat = NULL,
-                              lon = -110))
+  # error wrong product
+  expect_error(
+    mt_dates(
+      product = "MOD11A6",
+      lat = 40,
+      lon = -110
+    )
+  )
 
-  expect_is(mt_dates(product = "MOD11A2",
-                                site_id = mt_sites()$siteid[1]),
-            "data.frame")
+  # error missing latitude
+  expect_error(
+    mt_dates(
+      product = "MOD11A2",
+      lon = -110
+    )
+  )
 
-  expect_error(mt_dates(product = "MOD11A2",
-                     site_id = "test"))
+  expect_is(
+    mt_dates(
+      product = "MOD11A2",
+      site_id = mt_sites()$siteid[1]
+      ),
+  "data.frame"
+  )
+
+  expect_error(
+    mt_dates(
+      product = "MOD11A2",
+      site_id = "test"
+    )
+  )
 })
 
 test_that("test mt_subset()",{
 
-  expect_is(mt_subset(product = "MOD11A2",
-                    lat = 40,
-                    lon = -110,
-                    band = "LST_Day_1km",
-                    start = "2004-01-01",
-                    end = "2004-03-31"),
-            "data.frame")
+  # good query
+  expect_is(
+    mt_subset(
+      product = "MOD11A2",
+      lat = 40,
+      lon = -110,
+      band = "LST_Day_1km",
+      start = "2004-01-01",
+      end = "2004-03-31",
+      progress = FALSE
+    ),
+    "data.frame"
+  )
 
-  expect_error(mt_subset(product = "MOD11A2",
-                         lat = 40,
-                         band = "LST_Day_1km",
-                         start = "2004-01-01",
-                         end = "2004-03-31"))
+  # missing coordinate
+  expect_error(
+    mt_subset(
+      product = "MOD11A2",
+      lat = 40,
+      band = "LST_Day_1km",
+      start = "2004-01-01",
+      end = "2004-03-31"
+    )
+  )
 
-  expect_is(mt_subset(product = "MOD11A2",
-                                 site_id = "us_tennessee_neon_ornl",
-                                 band = "LST_Day_1km",
-                                 start = "2004-01-01",
-                                 end = "2004-03-31"),
-            "data.frame")
+  # no band provided
+  expect_error(mt_subset(
+    product = "MOD11A2",
+    lat = 40,
+    start = "2004-01-01",
+    end = "2004-03-31"
+  ))
 
-  expect_silent(mt_subset(product = "MOD11A2",
-                          lat = 40,
-                          lon = -110,
-                          band = "LST_Day_1km",
-                          start = "2004-01-01",
-                          end = "2004-03-31",
-                          internal = FALSE,
-                          progress = FALSE))
[…]

[…]
