+++ .travis.yml
@@ -17,4 +17,4 @@ r_packages:
   - covr
 
 after_success:
-- Rscript -e 'covr::codecov()'
+  - Rscript -e 'covr::codecov()'
