+++ README.md
@@ -76,6 +76,7 @@ expand)</summary>
 | km\_ab     | suppress the verbose output (integer)                                                                                           |
 | site\_name | a site identifier                                                                                                               |
 | site\_id   | a site\_id for predefined locations (not required)                                                                              |
+| progress   | logical, TRUE or FALSE (show download progress)                                                                                 |
 
 </p>
 
@@ -98,7 +99,7 @@ subset <- mt_subset(product = "MOD11A2",
                     internal = TRUE,
                     progress = FALSE)
 print(str(subset))
-#> Classes 'MODISTools' and 'data.frame':   36 obs. of  21 variables:
+#> 'data.frame':    36 obs. of  21 variables:
 #>  $ xllcorner    : chr  "-9370962.97" "-9370962.97" "-9370962.97" "-9370962.97" ...
 #>  $ yllcorner    : chr  "4446875.49" "4446875.49" "4446875.49" "4446875.49" ...
 #>  $ cellsize     : chr  "926.625433055834" "926.625433055834" "926.625433055834" "926.625433055834" ...
@@ -118,7 +119,7 @@ print(str(subset))
 #>  $ calendar_date: chr  "2004-01-01" "2004-01-09" "2004-01-17" "2004-01-25" ...
 #>  $ tile         : chr  "h09v05" "h09v05" "h09v05" "h09v05" ...
 #>  $ proc_date    : chr  "2015212185706" "2015212201022" "2015212213103" "2015213005429" ...
-#>  $ pixel        : chr  "1" "1" "1" "1" ...
+#>  $ pixel        : int  1 1 1 1 2 2 2 2 3 3 ...
 #>  $ value        : int  13135 13120 13350 13354 13123 13100 13324 13331 13098 13069 ...
 #> NULL
 ```
@@ -165,7 +166,7 @@ subsets <- mt_batch_subset(df = df,
                      end = "2004-02-01")
 
 print(str(subsets))
-#> Classes 'MODISTools' and 'data.frame':   8 obs. of  21 variables:
+#> 'data.frame':    8 obs. of  21 variables:
 #>  $ xllcorner    : chr  "-9370036.39" "-9370036.39" "-9370036.39" "-9370036.39" ...
 #>  $ yllcorner    : chr  "4447802.08" "4447802.08" "4447802.08" "4447802.08" ...
 #>  $ cellsize     : chr  "926.625433055834" "926.625433055834" "926.625433055834" "926.625433055834" ...
@@ -185,7 +186,7 @@ print(str(subsets))
 #>  $ calendar_date: chr  "2004-01-01" "2004-01-09" "2004-01-17" "2004-01-25" ...
 #>  $ tile         : chr  "h09v05" "h09v05" "h09v05" "h09v05" ...
 #>  $ proc_date    : chr  "2015212185706" "2015212201022" "2015212213103" "2015213005429" ...
-#>  $ pixel        : chr  "1" "1" "1" "1" ...
+#>  $ pixel        : int  1 1 1 1 1 1 1 1
 #>  $ value        : int  13098 13062 13297 13323 13098 13062 13297 13323
 #> NULL
 ```
