### 2018-10-31 73630b5
missing references + tests
