+++ R/mt_sites.R
@@ -5,7 +5,7 @@
 #' @return A data frame of all available MODIS Land Products Subsets
 #' pre-processed sites
 #' @keywords MODIS Land Products Subsets, products, meta-data
-#' @seealso \code{\link[MODISTools]{mt_product}}
+#' @seealso \code{\link[MODISTools]{mt_products}}
 #' \code{\link[MODISTools]{mt_bands}} \code{\link[MODISTools]{mt_dates}}
 #' @export
 #' @examples
