+++ vignettes/modistools-vignette.Rmd
@@ -20,7 +20,7 @@ library(MODISTools)
 
 ```
 
-The MODISTools package has as goal to facilitate the interface between R and the MODIS Land Product Subset API at the Oak Ridge National Laboratory DAAC. This programmatic interface to the ['MODIS Land Products Subsets' web services](https://modis.ornl.gov/data/modis_webservice.html) allows for easy downloads of 'MODIS' time series directly to your R workspace or your computer. Below an examples are provided on how to download a MODIS time series as well as list ancillary data.
+The MODISTools package has as goal to facilitate the interface between R and the MODIS Land Product Subset API at the Oak Ridge National Laboratory DAAC. This programmatic interface to the ['MODIS Land Products Subsets' web services](https://modis.ornl.gov/data/modis_webservice.html) allows for easy downloads of 'MODIS' time series directly to your R workspace or your computer. Below an example is provided on how to download a MODIS time series as well as list ancillary data.
 
 ### Listing products / bands / dates
 
