### 2018-10-31 ce75e3f
update tests + parameter checks
