### 2018-10-31 f4e7044
more test routines
