### 2018-10-31 9251ab2
remove buttons
