### 2018-11-16 #246 @karthik
@etiennebr Thanks for the thoughtful response. I think the next steps would be to see if @khufkens has adequately addressed the issues that you've raised. Otherwise you can point out the ones that remain unresolved.  If it's easier, you can also open issues on his repo and tag `ropensci/onboarding#246` and they will appear as a list here (with automatic updates on when those issues are closed).
