### 2018-12-29 khufkens/MODISTools#10 @etiennebr
Conversion from sinusoidal lower-left corner to lon lat
I suspect there is something wrong with the conversion from sinusoidal to angular. I'm not familiar with the sinusoidal projection, but I would suspect that the conversion using a constant is inacurate.  https://github.com/khufkens/MODISTools/blob/1adc07bd8572454bf52ba629c7cf686c60e1398b/R/coordinate_conversion.R#L108

I think a safer approach would be to build the footprint from longitude latitude. Otherwise having a few tests with different locations across latitudes would convince me :)

It would also be good to document the origin of the MODIS coordinate system parameters.
https://github.com/khufkens/MODISTools/blob/1adc07bd8572454bf52ba629c7cf686c60e1398b/R/coordinate_conversion.R#L36-L40

ropensci/onboarding#246
