+++ R/coordinate_conversion.R
@@ -50,14 +50,17 @@ sin_to_ll <- function(x, y){
   return(coords)
 }
 
-#' Convert lat / lon coordinates to sf bounding box polygon
+#' Converts lower-left sinusoidal coordinates to lat-lon sf bounding box
 #'
-#' @param lat latitude generated by \code{\link[MODISTools]{sin_to_ll}}
-#' @param lon longitude generated by \code{\link[MODISTools]{sin_to_ll}}
-#' @param cell_size cell size provided by MODISTools
-#' @param nrows cell size provided by MODISTools
-#' @param ncols cell size provided by MODISTools
-#' @seealso \code{\link[MODISTools]{sin_to_ll}}
+#' @param xllcorner lower left x coordinate as provided by
+#' \code{\link[MODISTools]{mt_subset}}
+#' @param yllcorner lower left y coordinate as provided by
+#' \code{\link[MODISTools]{mt_subset}}
+#' @param cellsize cell size provided by \code{\link[MODISTools]{mt_subset}}
+#' @param nrows cell size provided by \code{\link[MODISTools]{mt_subset}}
+#' @param ncols cell size provided by \code{\link[MODISTools]{mt_subset}}
+#' @seealso \code{\link[MODISTools]{sin_to_ll}},
+#' \code{\link[MODISTools]{mt_subset}}
 #' @keywords MODIS Land Products Subsets, products, meta-data
 #' @export
 #' @examples
@@ -80,9 +83,9 @@ sin_to_ll <- function(x, y){
 #'
 #' # convert to bounding box
 #' bb <- apply(subset, 1, function(x){
-#'   ll_to_bb(lon = x['longitude_ll'],
-#'            lat = x['latitude_ll'],
-#'            cell_size = x['cellsize'],
+#'   mt_bbox(lon = x['xllcorner'],
+#'            lat = x['yllcorner'],
+#'            cellsize = x['cellsize'],
 #'            nrows = x['nrows'],
 #'            ncols = x['ncols'])
 #' })
@@ -90,22 +93,20 @@ sin_to_ll <- function(x, y){
 #' head(bb)
 #'}
 
-ll_to_bb <- function(
-  lat,
-  lon,
-  cell_size,
+mt_bbox <- function(
+  xllcorner,
+  yllcorner,
+  cellsize,
   nrows,
   ncols
   ){
 
   # check parameters
-  if(missing(cell_size) | missing(lat) | missing(lon) | missing(nrows) |
+  if(missing(cellsize) | missing(yllcorner) |
+     missing(xllcorner) | missing(nrows) |
      missing(ncols)){
     stop("missing parameter, all parameters are required")
   }
-
-  # meters to a degree
-  m_deg <- (0.00001/1.1132)
 
   # transform matrix
   trans <- c(
@@ -117,16 +118,19 @@ ll_to_bb <- function(
   )
 
   # create coordinate matrix
-  m <- c(-as.numeric(cell_size) * as.numeric(ncols) * m_deg,
-         as.numeric(cell_size) * as.numeric(nrows) * m_deg) * trans +
-    c(as.numeric(lon), as.numeric(lat))
+  m <- c(as.numeric(cellsize) * as.numeric(ncols),
+         as.numeric(cellsize) * as.numeric(nrows)) * trans +
+    c(as.numeric(xllcorner), as.numeric(yllcorner))
   m <- matrix(m, 5, 2, byrow = TRUE)
 
   # convert to a sf polygon, with the proper
   # projection etc.
   p <- sf::st_linestring(m)
   p <- sf::st_cast(p, "POLYGON")
-  p <- sf::st_sfc(p, crs = 4326)
[…]
