+++ man/mt_bbox.Rd
@@ -1,24 +1,26 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/coordinate_conversion.R
-\name{ll_to_bb}
-\alias{ll_to_bb}
-\title{Convert lat / lon coordinates to sf bounding box polygon}
+\name{mt_bbox}
+\alias{mt_bbox}
+\title{Converts lower-left sinusoidal coordinates to lat-lon sf bounding box}
 \usage{
-ll_to_bb(lat, lon, cell_size, nrows, ncols)
+mt_bbox(xllcorner, yllcorner, cellsize, nrows, ncols)
 }
 \arguments{
-\item{lat}{latitude generated by \code{\link[MODISTools]{sin_to_ll}}}
+\item{xllcorner}{lower left x coordinate as provided by
+\code{\link[MODISTools]{mt_subset}}}
 
-\item{lon}{longitude generated by \code{\link[MODISTools]{sin_to_ll}}}
+\item{yllcorner}{lower left y coordinate as provided by
+\code{\link[MODISTools]{mt_subset}}}
 
-\item{cell_size}{cell size provided by MODISTools}
+\item{cellsize}{cell size provided by \code{\link[MODISTools]{mt_subset}}}
 
-\item{nrows}{cell size provided by MODISTools}
+\item{nrows}{cell size provided by \code{\link[MODISTools]{mt_subset}}}
 
-\item{ncols}{cell size provided by MODISTools}
+\item{ncols}{cell size provided by \code{\link[MODISTools]{mt_subset}}}
 }
 \description{
-Convert lat / lon coordinates to sf bounding box polygon
+Converts lower-left sinusoidal coordinates to lat-lon sf bounding box
 }
 \examples{
 
@@ -40,9 +42,9 @@ subset <- cbind(subset, lat_lon)
 
 # convert to bounding box
 bb <- apply(subset, 1, function(x){
-  ll_to_bb(lon = x['longitude_ll'],
-           lat = x['latitude_ll'],
-           cell_size = x['cellsize'],
+  mt_bbox(lon = x['xllcorner'],
+           lat = x['yllcorner'],
+           cellsize = x['cellsize'],
            nrows = x['nrows'],
            ncols = x['ncols'])
 })
@@ -51,7 +53,8 @@ head(bb)
 }
 }
 \seealso{
-\code{\link[MODISTools]{sin_to_ll}}
+\code{\link[MODISTools]{sin_to_ll}},
+\code{\link[MODISTools]{mt_subset}}
 }
 \keyword{Land}
 \keyword{MODIS}
