### 2018-12-30 3f9f252
typo + docs
