### 2018-12-30 ae59101
docs + unit testing changes
