+++ man/mt_bbox.Rd
@@ -42,11 +42,11 @@ subset <- cbind(subset, lat_lon)
 
 # convert to bounding box
 bb <- apply(subset, 1, function(x){
-  mt_bbox(lon = x['xllcorner'],
-           lat = x['yllcorner'],
-           cellsize = x['cellsize'],
-           nrows = x['nrows'],
-           ncols = x['ncols'])
+  mt_bbox(xllcorner = x['xllcorner'],
+          yllcorner = x['yllcorner'],
+          cellsize = x['cellsize'],
+          nrows = x['nrows'],
+          ncols = x['ncols'])
 })
 
 head(bb)
