### 2018-12-30 khufkens/MODISTools#10 @khufkens
I've solved both issues after some detective work. Since the cellsize column doesn't change with changing latitude I think it is safe to assume that this is the grid size of the original SIN grid, for which the lower left coordinate is provided. Combining both I can then directly calculate the bounding box in SIN and convert to lat/lon.

https://github.com/khufkens/MODISTools/blob/088b222f82ddce7193a54a3ea8ba68214a7639cf/R/coordinate_conversion.R#L130

This is consistent with taking subsets as x rows and y columns from the original center pixel, rather than doing it region based (overlay of pixels), consistent with the km parameters only taking integers not floats if it was truly freeform. However, this also changes the interpretation of what is selected in lat/lon (basically the km above-below / left-right are note km in lat/lon).
