### 2018-12-30 c09b9c1
additional information on the sinusoidal projection
