### 2019-01-04 khufkens/MODISTools#10 @khufkens
This is resolved imho, will re-open if disputed.
