### 2019-01-04 c7a7d94
return NULL upon empty data query, not to crash the batch routine for invalid locations
