### 2019-01-04 8ca2abc
trapping capitalization errors in the column names
