+++ R/mt_subset.R
@@ -124,8 +124,7 @@ mt_subset <- function(
 
   # check if something remains
   if (nrow(dates) == 0){
-    warning("No data points exist for the selected date range...")
-    return(invisible(NULL))
+    stop("No data points exist for the selected date range...")
   }
 
   # list breaks, for downloads in chunks
