### 2019-01-08 9ca600c
resolving goodpractice NOTE on memoise not being used with explicit @ImportFrom statement
