### 2019-01-08 #246 @khufkens
Thanks for the vote of confidence @etiennebr. I acknowledge that `sf` support would be nice, but I also want to make sure that there is something for people to work with. Since the CRAN (re)release the package has had some good traffic. So the demand for the tool is there.

New features can be added incrementally without breaking things. I think with the feedback the package is also better structured which makes all the difference when moving forward.
