### 2019-01-26 #246 @khufkens
Thanks. Take your time, no rush. Previous version is still up and things only got better.
