+++ cran-comments.md
@@ -3,8 +3,9 @@ http://cran.r-project.org/web/packages/policies.html
 
 ## test environments, local, CI and r-hub
 
-- local OSX / Ubuntu 16.04 install on R 3.4.4
+- local OSX / Ubuntu 16.04 install on R 3.5
 - Ubuntu 14.04 on Travis-CI (devel / release)
+- rhub check on windows OK
 - codecove.io code coverage at ~90%
 
 ## local / Travis CI R CMD check results
