### 2019-03-14 6ea4b36
Delete testthat.r
