+++ tests/testthat.r
@@ -1,6 +1,0 @@
-# Run all unit tests
-
-# This is a wrapper around all the tests specified
-# in the tests/testthat directory. These tests are visible
-# to users and serve as small examples as well.
-testthat::test_check("MODISTools")
