### 2019-03-14 6fbd6b1
mt_to_raster() function
