+++ NAMESPACE
@@ -7,5 +7,6 @@ export(mt_dates)
 export(mt_products)
 export(mt_sites)
 export(mt_subset)
+export(mt_to_raster)
 export(sin_to_ll)
 importFrom(memoise,memoise)
