### 2019-03-14 fba5836
adding raster requirement
