+++ R/mt_subset.R
@@ -4,7 +4,7 @@
 #' at a particular location.
 #'
 #' @param product a valid MODIS product name
-#' @param band band to download
+#' @param band band or bands (as a character vector) to download
 #' @param lat latitude in decimal degrees
 #' @param lon longitude in decimal degrees
 #' @param start start date
@@ -86,7 +86,7 @@ mt_subset <- function(
   bands <- mt_bands(product)
 
   # error trap band
-  if (missing(band) | !(band %in% bands$band) ){
+  if (missing(band) | !all(band %in% bands$band) ){
     stop("please specify a band, or check your product band combination ...")
   }
 
@@ -130,145 +130,158 @@ mt_subset <- function(
   # list breaks, for downloads in chunks
   breaks <- seq(1, nrow(dates), 10)
 
-  # start progress bar chuncks
-  if(progress){
-    message("Downloading chunks:")
-    env <- environment()
-    counter <- 0
-    pb <- utils::txtProgressBar(
-      min = 0,
-      max = length(breaks),
-      style = 3
-      )
-  }
+  # grab site name
+  site <- ifelse(missing(site_id), site_name, site_id)
 
-  # loop over all 10 value breaks
-  subset_data <- lapply(breaks, function(b){
+  # create a complete list of the data
+  complete_data <- lapply(band, function(band_name){
 
-    # grab last date for subset
-    if(b == breaks[length(breaks)]){
-      end_date <-dates$modis_date[nrow(dates)]
-    } else {
-      end_date <- dates$modis_date[b+9]
-    }
-
-    # construct the query to be served to the server
-    query <- list("latitude" = lat,
-                  "longitude" = lon,
-                  "band" = band,
-                  "startDate" = dates$modis_date[b],
-                  "endDate" = end_date,
-                  "kmAboveBelow" = km_ab,
-                  "kmLeftRight" = km_lr)
-
-    # try to download the data
-    json_chunk <- httr::GET(url = url,
-                         query = query,
-                         httr::write_memory())
-
-    # trap errors on download, return a detailed
-    # API error statement
-    if (httr::http_error(json_chunk)){
-      warning(httr::content(json_chunk), call. = FALSE)
-      return(NULL)
-    }
-
-    # check the content of the response
-    if (httr::http_type(json_chunk) != "application/json") {
-      warning("API did not return json...", call. = FALSE)
-      return(NULL)
-    }
-
-    # grab content from cached json chunk
-    chunk <- jsonlite::fromJSON(httr::content(json_chunk, "text",
-                                              encoding = "UTF-8"),
-                                simplifyVector = TRUE)
-
-    # set progress bar
+    # start progress bar chuncks
     if(progress){
-      tmp <- get("counter", envir = env)
-      assign("counter", tmp + 1 ,envir = env)
-      utils::setTxtProgressBar(get("pb", envir = env), tmp + 1)
+      message("Downloading chunks:")
+      env <- environment()
+      counter <- 0
+      pb <- utils::txtProgressBar(
+        min = 0,
+        max = length(breaks),
+        style = 3
+        )
     }
 
-    # return data
-    return(chunk)
+
+    # loop over all 10 value breaks
+    subset_data <- lapply(breaks, function(b){
+
+      # grab last date for subset
+      if(b == breaks[length(breaks)]){
+        end_date <-dates$modis_date[nrow(dates)]
+      } else {
+        end_date <- dates$modis_date[b+9]
+      }
+
+      # construct the query to be served to the server
+      query <- list("latitude" = lat,
+                    "longitude" = lon,
+                    "band" = band_name,
+                    "startDate" = dates$modis_date[b],
+                    "endDate" = end_date,
+                    "kmAboveBelow" = km_ab,
[…]
