### 2019-04-21 4ca7e3a
skip tests on cran
