+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: MODISTools
 Title: Interface to the 'MODIS Land Products Subsets' Web Services
-Version: 1.1.0
+Version: 1.1.1
 Authors@R: person("Hufkens","Koen",
 		  email="koen.hufkens@gmail.com",
 		  role=c("aut", "cre"),
@@ -20,7 +20,8 @@ Imports:
     raster,
     stats,
     memoise,
-    jsonlite
+    jsonlite,
+    methods
 License: AGPL-3
 LazyData: true
 ByteCompile: true
