### 2019-04-22 e090c62
adding error traps and unit checks mt_to_raster()
