+++ NAMESPACE
@@ -10,3 +10,4 @@ export(mt_subset)
 export(mt_to_raster)
 export(sin_to_ll)
 importFrom(memoise,memoise)
+importFrom(raster,stack)
