### 2019-05-08 faa7862
adding documentation changes and error traps to `mt_to_raster()`
