+++ man/mt_subset.Rd
@@ -43,8 +43,10 @@ mt_subset(product, band, lat, lon, start = "2000-01-01",
 A data frame combining meta-data and actual data values.
 }
 \description{
-Lists all available dates for a MODIS Land Products Subset product
-at a particular location.
+Download a MODIS Land Products Subset product
+for a given point location buffered with a given amount of kilometers
+left-right, top-bottom for a given location (provided as latitude and
+longitude values).
 }
 \examples{
 
