### 2019-05-08 1a64e00
adding rgdal as required packages to render vignette correctly
