### 2019-05-08 3f20345
sorting out rgal requirement for vignette building
