+++ R/mt_to_raster.R
@@ -3,7 +3,7 @@
 #' Convert tidy MODISTools data to a raster (stack)
 #'
 #' @param df a valid MODISTools data frame with a single band (filter for a
-#' particular band using the dplyr \code{filter()} function or base \code{subset()})
+#' particular band using the dplyr \code{filter()} function or base \code{subset()}
 #' @param reproject reproject output to lat / long (default = \code{FALSE})
 #' @return A raster stack populated with the tidy dataframe values
 #' @keywords MODIS Land Products Subsets, products
