+++ R/data.R
@@ -1,0 +1,15 @@
+#' arcachon_lai
+#'
+#' MODIS leaf area index (LAI) around the French town of Arcachon
+#' derived from the MODIS MOD15A2H product (band Lai_500m).
+#'
+#' @format A MODISTools tidy data frame
+"arcachon_lai"
+
+#' arcachon_lc
+#'
+#' MODIS land cover (IGBP) around the French town of Arcachon
+#' derived from the MODIS MCD12Q2 product (band LC_Type1).
+#'
+#' @format A MODISTools tidy data frame
+"arcachon_lc"
