+++ vignettes/modistools-vignette.Rmd
@@ -19,6 +19,7 @@ knitr::opts_chunk$set(
 library(MODISTools)
 library(raster)
 library(rgdal)
+library(tidyverse)
 ```
 
 The MODISTools package has as goal to facilitate the interface between R and the MODIS Land Product Subset API at the Oak Ridge National Laboratory Distributed Active Archive Center (DAAC). This programmatic interface to the ['MODIS Land Products Subsets' web services](https://modis.ornl.gov/data/modis_webservice.html) allows for easy downloads of 'MODIS' time series (of single pixels or small regions of interest) directly to your R workspace or your computer. Below an example is provided on how to download a MODIS time series as well as list ancillary data.
@@ -52,25 +53,48 @@ head(dates)
 
 # Downloading MODIS time series
 
-Once you decide on which data to download using the above functions you can use these parameters to download a time series using the `mt_subset()` function. The below query downloads MOD13Q1 based normalized vegetation index (NDVI) for the year 2014. The location is named 'testsite'. The output is saved to a variable called subset in the R workspace (as defined by the parameter internal = TRUE, when set to FALSE the data is written to file).
+Once you decide on which data to download using the above functions you can use these parameters to download a time series using the `mt_subset()` function. The below query downloads MOD15A2H based leaf area index (LAI) data for the year 2014 for an area around the Arcachon basin in the south west of France. We will also download land cover data (MCD12Q1, IGBP) at a similar scale. The location is named 'arcachon'. The output is saved to a variables called subset and LC in the R workspace (as defined by the parameter internal = TRUE, when set to FALSE the data is written to file). Keep in mind that this operation might take a while.
 
-```{r eval = TRUE}
-# download data
-subset <- mt_subset(product = "MOD13Q1",
-                    lat = 42.534171,
-                    lon = -72.179003,
-                    band = "250m_16_days_NDVI",
+```{r eval = FALSE}
+# download the MODIS land cover (IGBP) and NDVI data
+# for a region around the French city and basin of Arcachon
+arcachon_lai <- mt_subset(product = "MOD15A2H",
+                    lat = 44.656286,
+                    lon =  -1.174748,
+                    band = "Lai_500m",
                     start = "2004-01-01",
                     end = "2004-12-30",
-                    km_lr = 0,
-                    km_ab = 0,
-                    site_name = "testsite",
+                    km_lr = 20,
+                    km_ab = 20,
+                    site_name = "arcachon",
                     internal = TRUE,
                     progress = FALSE)
-head(subset)
+
+arcachon_lc <- mt_subset(product = "MCD12Q1",
+  lat = 44.656286,
+  lon =  -1.174748,
+  band = "LC_Type1",
+  start = "2004-01-01",
+  end = "2004-3-20",
+  km_lr = 20,
+  km_ab = 20,
+  site_name = "arcachon",
+  internal = TRUE,
+  progress = FALSE)
 ```
 
 The output format is a *tidy* data frame, as shown above. When witten to a csv with the parameter `internal = FALSE` this will result in a flat file on disk.
+
+```{r echo = FALSE}
+# load data
+data("arcachon_lai")
+data("arcachon_lc")
+```
+
+```{r}
+head(arcachon_lai)
+head(arcachon_lc)
+```
 
[…]
