+++ man/arcachon_lc.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{arcachon_lc}
+\alias{arcachon_lc}
+\title{arcachon_lc}
+\format{A MODISTools tidy data frame}
+\usage{
+arcachon_lc
+}
+\description{
+MODIS land cover (IGBP) around the French town of Arcachon
+derived from the MODIS MCD12Q2 product (band LC_Type1).
+}
+\keyword{datasets}
