### 2019-05-08 7631c60
dplyr requirement for data wrangling in vignette
