### 2019-05-08 3d74f68
more limited library requirements
