### 2019-05-17 #246 @pmnatural
#### Final approval (post-review)

Estimated hours spent reviewing:
