+++ .travis.yml
@@ -4,8 +4,10 @@ cache: packages
 warnings_are_errors: true
 
 before_install:
-  - sudo apt-get install --yes udunits-bin libudunits2-dev
-  - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev libgdal1-dev
+  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
+  - sudo apt-get update
+  - sudo apt-get install --yes libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
+  - sudo apt-get install --yes udunits-bin
   - sudo apt-get install --yes libnetcdf-dev netcdf-bin
 
 r:
