### 2019-05-18 2c6fb4a
force continuation of install
