+++ .travis.yml
@@ -4,8 +4,8 @@ cache: packages
 warnings_are_errors: true
 
 before_install:
-  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable
-  - sudo apt-get update
+  - sudo add-apt-repository --yes ppa:ubuntugis/ubuntugis-unstable
+  - sudo apt-get update --yes
   - sudo apt-get install --yes libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
   - sudo apt-get install --yes udunits-bin
   - sudo apt-get install --yes libnetcdf-dev netcdf-bin
