+++ .travis.yml
@@ -3,11 +3,21 @@ sudo: required
 cache: packages
 warnings_are_errors: true
 
-before_install:
-  - sudo apt-get install --yes libudunits2-dev libgdal-dev libgeos-dev libproj-dev 
-  - sudo apt-get install --yes udunits-bin
-  - sudo apt-get install --yes libnetcdf-dev netcdf-bin
-
+addons:
+  apt:
+    sources:
+      - sourceline: 'ppa:ubuntugis/ubuntugis-unstable'
+    packages:
+      - libudunits2-dev
+      - libunits-bin
+      - libproj-dev
+      - libgeos-dev
+      - libspatialite-dev
+      - libgdal-dev
+      - libjson-c-dev
+      - libnetcdf-dev
+      - netcdf-bin
+      
 r:
   - release
   - devel
