### 2019-07-12 #246 @khufkens
Hi @karthik I was notified by someone that the website link of the package is broken. I forgot to change the link to an ropensci account one.

It currently reads (which leads nowhere):
https://khufkens.github.io/MODISTools/

but should be
https://ropensci.github.io/MODISTools/

Is there someone at your end who can update this so that people can access examples etc?
